+++ src/main/java/shoppinglist/patches/StorePotionPurchasePatch.java
@@ -23,9 +23,11 @@ public class StorePotionPurchasePatch {
         // Intercept the interaction if the item is intended to be added to the list.
         if (KeyHelper.isAltPressed()) {
             // Add the item to the list.
-            ShoppingListPanelField.shoppingList.get(AbstractDungeon.shopScreen).addItem(storePotion);
+            panel.addItem(storePotion);
 
             return SpireReturn.Return(null);
+        } else if (AbstractDungeon.player.gold >= storePotion.price) {
+            panel.removeItem(storePotion);
         }
 
         return SpireReturn.Continue();